services/app/requirements.txt
mysqlclient==2.2.4
SQLAlchemy==2.0.35
python-dotenv==1.0.1
imapclient==3.0.0
beautifulsoup4==4.12.3
openai==1.43.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2